xiaomi/requirements.txt
prometheus-client==0.8.0
pycrypto==2.6.1
requests==2.24.0

